(fix) Se corrigió no haber deshabilitado el botón de eliminar para los usuarios no administradores.

+++ vistas/usuarios.py
@@ -129,6 +129,9 @@ def BtnEliminar(usuario: str, set_usuarios: Callable[[Any], None]):
     sesion = contexto["sesion"]
 
     async def eliminar(_):
+        if sesion["rol"] != "admin":
+            return
+
         await eliminar_usuario(usuario)
         set_usuarios(
             lambda anteriores: list(
